==> R/09_valente_2019.R <==
# <><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><>
# AUTHOR:       Philippe Massicotte
#
# DESCRIPTION:  Explore the data published in xxx to make sure that COASTLOOC
# data is not present.
# <><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><>

rm(list = ls())

df <- data.table::fread(here("data/raw/valente_2019/Orginal_files_including_metadata/insitudb_iopskdtsm_2019-02.csv")) %>%
  as_tibble() %>%
  distinct(time, lat, lon)

df

df_sf <- df %>%
  st_as_sf(coords = c("lon", "lat"), crs = 4326)

df_sf %>%
  ggplot() +
  geom_sf()

bbox_4326 <- st_bbox(c(
  xmin = -20,
  xmax = 25,
  ymax = 25,
  ymin = 60
), crs = st_crs(4326))

df_sf <- df_sf %>%
  st_crop(bbox_4326)

df_sf %>%
  ggplot() +
  geom_sf()

ne_land <-
  rnaturalearth::ne_download(
    category = "physical",
    type = "land",
    returnclass = "sf",
    scale = "large"
  ) %>%
  st_crop(bbox_4326)

map_crs <- 3035

# Plot --------------------------------------------------------------------

p <- df_sf %>%
  ggplot() +
  geom_sf(data = ne_land, size = 0.1) +
  geom_sf(size = 0.5) +
  coord_sf(
    xlim = c(1680000, 5100000),
    ylim = c(979000, 3750000),
    expand = FALSE,
    crs = map_crs
  ) +
  facet_wrap(~lubridate::year(time)) +
  labs(
    title = "Map with Valente et al. data",
    subtitle = "There are no overlaps with the COASTLOOC data."
  )

ggsave(
  here("graphs/09_map_valente_et_al.pdf"),
  device = cairo_pdf,
  width = 8,
  height = 8
)

knitr::plot_crop(here("graphs/09_map_valente_et_al.pdf"))
